requirements.txt
fastapi==0.115.0
uvicorn==0.30.0

httpx==0.27.0
anyio==4.4.0

pydantic==2.9.0
pydantic-settings==2.4.0
email-validator==2.2.0

structlog==24.1.0

PyYAML==6.0.2

python-dateutil==2.9.0.post0

# Dev/Test (optional but recommended)
pytest==8.2.0
pytest-asyncio==0.23.6
